+++ src/environments/environment.ts
@@ -3,8 +3,8 @@
 // The list of file replacements can be found in `angular.json`.
 
 const devUrl = 'https://dev-api.noplaceleft.tools/api/';
-const localUrl = 'http://localhost:9000/api/';
-// const localUrl = devUrl;
+// const localUrl = 'http://localhost:9000/api/';
+const localUrl = devUrl;
 
 export const environment = {
     production: false,